MCPFromScratch/client/requirements.txt
websockets>=12.0
pydantic
python-dotenv
openai
